requirements.txt
git+https://github.com/Rapptz/discord.py.git
jishaku
discord-ext-menus
psutil==5.9.8
requests
aiosqlite
gtts
langdetect
python-dotenv
PyYAML
openai
duckduckgo-search==6.3.6
pillow>=10.2.0
colorama
chess
groq
sympy
aiohttp
pytz
tabulate
beautifulsoup4
aiofiles
python-dateutil
tenacity
httpx
typing-extensions
google-generativeai
